Replace old types with new files in project
