--- ReactiveCocoa/Swift/NewAction.swift
@@ -1,42 +1,0 @@
-/// Represents an action that will perform side effects or a transformation when
-/// executed with an input.
-///
-/// Actions enforce serial execution. Any attempt to execute an action multiple
-/// times concurrently will return an error.
-public final class Action<Input, Output> {
-	/// A signal of all values generated from applications of the Action.
-	///
-	/// In other words, this will send every value from every signal generated
-	/// by each SignalProducer returned from apply().
-	public let values: Signal<Output>
-
-	/// A signal of all errors generated from applications of the Action,
-	/// _except_ `RACAction.ActionNotEnabled` errors.
-	///
-	/// In other words, this will send errors from every signal generated by
-	/// each SignalProducer returned from apply().
-	public let errors: Signal<NSError>
-
-	/// Whether the action is currently executing.
-	public let executing: Property<Bool>
-
-	/// Whether the action is currently enabled.
-	public let enabled: Property<Bool>
-
-	/// Initializes an action that will be conditionally enabled, and create a
-	/// SignalProducer for each input.
-	public init(enabledIf: Property<Bool>, _ execute: Input -> SignalProducer<Output>)
-
-	/// Initializes an action that will be enabled by default, and create a
-	/// SignalProducer for each input.
-	public convenience init(_ execute: Input -> SignalProducer<Output>)
-
-	/// Creates a SignalProducer that, when started, will execute the action
-	/// with the given input, then forward the results upon the produced Signal.
-	///
-	/// If the action is disabled when the returned SignalProducer is started,
-	/// the produced signal will send an `NSError` corresponding to
-	/// `RACError.ActionNotEnabled`, and nothing will be sent upon `values` or
-	/// `errors` for that particular signal.
-	public func apply(input: Input) -> SignalProducer<Output>
-}
